renames commands
