+++ src/Core/Domain/Product/Supplier/Command/UpdateProductSuppliersCommand.php
@@ -34,10 +34,9 @@ use PrestaShop\PrestaShop\Core\Domain\Product\ValueObject\ProductId;
 use PrestaShop\PrestaShop\Core\Domain\Supplier\Exception\SupplierException;
 
 /**
- * @todo: I have also different behavior when combinations are available.
  * Updates supplier data related with specific product.
  */
-class EditProductSupplierCommand
+class UpdateProductSuppliersCommand
 {
     /**
      * @var Supplier[]
